+++ src/App.tsx
@@ -1,13 +1,15 @@
 import React from "react";
-import Dashboard from "./Views/Dashboard/Dashboard";
-import {Client} from "reduct-js";
-import {Image, Layout, Menu} from "antd";
+import {Divider, Image, Layout, Menu} from "antd";
+import {RouteComponentProps, withRouter} from "react-router-dom";
 
 import logo from "./main_logo.png";
 import "antd/dist/antd.variable.min.css";
 import "./App.css";
 
 import {ConfigProvider} from "antd";
+import {IBackendAPI} from "./BackendAPI";
+import {Routes} from "./Components/Routes";
+import {BorderOuterOutlined, LogoutOutlined} from "@ant-design/icons";
 
 ConfigProvider.config({
     theme: {
@@ -16,32 +18,74 @@ ConfigProvider.config({
 });
 
 
-function App() {
-    let url = process.env.REACT_APP_STORAGE_URL;
-    if (url === undefined) {
-        let path = window.location.pathname;
-        path = path.replace(process.env.PUBLIC_URL, "");
-        url = `${window.location.protocol}//${window.location.host}${path}`;
+interface Props extends RouteComponentProps {
+    backendApi: IBackendAPI;
+}
+
+type State = {
+    isAllowed?: boolean;
+}
+
+class App extends React.Component<Props, State> {
+    constructor(props: Props) {
+        super(props);
+        this.state = {};
     }
 
-    const client = new Client(url);
-    return (
-        <div className="App">
+    componentDidMount() {
+        this.props.backendApi.isAllowed()
+            .then((isAllowed) => this.setState({isAllowed}))
+            .catch((err) => console.error(err));
+    }
+
+    render() {
+        if (this.state.isAllowed === undefined) {
+            return <div/>;
+        }
+
+        const {backendApi, history} = this.props;
+        const onLogout = async () => {
+            backendApi.logout();
+            this.setState({isAllowed: false});
+        };
+
+        const onLogin = async () => {
+            this.setState({isAllowed: true});
+            this.props.history.push("/");
+        };
+
+        console.log(this.state.isAllowed);
+
+        return <div className="App">
             <Layout>
                 <Layout.Sider className="Sider">
-                    <Menu className="MenuItem">
-                        <a href="https://reduct-storage.dev">
+                    <Menu className="MenuItem" selectable={false} triggerSubMenuAction="click">
+                        <a href="https://reduct-storage.dev" title="https://reduct-storage.dev">
                             <Image src={logo} preview={false}/>
                         </a>
+
+                        <Divider/>
+                        {this.state.isAllowed ? <>
+                            <Menu.Item icon={<BorderOuterOutlined/>} onClick={() => history.push("/dashboard")}>
+                                Dashboard
+                            </Menu.Item>
+
+
+                            <Menu.Item onClick={onLogout} icon={<LogoutOutlined/>}>
+                                Logout
+                            </Menu.Item> </> :
+                            <div/>
+                        }
                     </Menu>
                 </Layout.Sider>
                 <Layout.Content>
-                    <Dashboard client={client}/>
+                    <Routes {...this.props} isAllowed={this.state.isAllowed} onLogin={onLogin}/>
                 </Layout.Content>
 
             </Layout>
-        </div>
-    );
+        </div>;
+    }
 }
 
-export default App;
+// @ts-ignore
+export default withRouter(App);