+++ src/Helpers/TestHelpers.tsx
@@ -1,3 +1,23 @@
+import {createMemoryHistory} from "history";
+import {RouteComponentProps} from "react-router-dom";
+
+export const makeRouteProps = (): RouteComponentProps => {
+    return {
+        match: {
+            isExact: false,
+            path: "",
+            url: "",
+            params: {id: "1"}
+        },
+            // @ts-ignore
+        location: {
+
+        },
+        // @ts-ignore
+        history: createMemoryHistory()
+    };
+};
+
 export const mockJSDOM = () => {
     Object.defineProperty(window, "matchMedia", {
         writable: true,