+++ src/Views/Dashboard/Dashboard.test.tsx
@@ -1,11 +1,10 @@
 import React from "react";
 import {mount, ReactWrapper} from "enzyme";
 import waitUntil from "async-wait-until";
-import {mockJSDOM} from "../../mockJSDOM";
+import {mockJSDOM} from "../../Helpers/TestHelpers";
 
 import Dashboard from "./Dashboard";
 import {Client, ServerInfo} from "reduct-js";
-
 
 
 describe("Dashboard", () => {
@@ -15,6 +14,12 @@ describe("Dashboard", () => {
     });
 
     const client = new Client("");
+    const backend = {
+        get client() { return client; },
+        login: jest.fn(),
+        logout: jest.fn(),
+        isAllowed: jest.fn(),
+    };
 
     const waitFor = async (wrapper: ReactWrapper, selector: string) => {
         await waitUntil(() => wrapper.update().find(selector).hostNodes().length > 0);
@@ -38,7 +43,7 @@ describe("Dashboard", () => {
         client.getInfo = jest.fn().mockResolvedValue(serverInfo);
         client.getBucketList = jest.fn().mockResolvedValue([]);
 
-        const wrapper = mount(<Dashboard client={client}/>);
+        const wrapper = mount(<Dashboard backendApi={backend}/>);
         const html = await waitFor(wrapper, "#ServerInfo");
 
         expect(html.text()).toContain("0.4.0");
@@ -57,7 +62,7 @@ describe("Dashboard", () => {
         }
         ]);
 
-        const wrapper = mount(<Dashboard client={client}/>);
+        const wrapper = mount(<Dashboard backendApi={backend}/>);
         const html = await waitFor(wrapper, "#bucket_1");
         expect(html.text()).toContain("bucket_1");
         expect(html.text()).toContain("1 KB");