+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Use Reduct's colours for UI, [PR-7](https://github.com/reduct-storage/web-console/pull/7)
 * Implement changing bucket settings, [PR-9](https://github.com/reduct-storage/web-console/pull/9)
 * Show error in CreateOrUpdate form if don't get settings, [PR-10](https://github.com/reduct-storage/web-console/pull/10)
+* Add Login Form for authentication with token, [PR0-14](https://github.com/reduct-storage/web-console/pull/14)
 
 **Bugs**: